refactored upgrade. moved code to upgrade_server
moved upgrade related code from cluster_server to upgrade_server

fixed basic_server_ops

fix
